[QAE1136] Case Summary PDFs - sceleton
