+++ app/pdf_generators/report_pdf_form_answer_pointer_base.rb
@@ -1,0 +1,35 @@
+require "prawn/measurement_extensions"
+
+class ReportPdfFormAnswerPointerBase
+  include SharedPdfHelpers::DrawElements
+  include SharedPdfHelpers::DataHelpers
+  include FormAnswersBasePointer
+
+  UNDEFINED_VALUE = "Not filled yet..."
+
+  attr_reader :user,
+              :form_answer,
+              :award_form,
+              :all_questions,
+              :answers,
+              :filled_answers,
+              :data,
+              :pdf_doc
+
+  def initialize(pdf_doc, form_answer)
+    @pdf_doc = pdf_doc
+    @form_answer = form_answer
+    @user = form_answer.user
+    @answers = fetch_answers
+    @award_form = form_answer.award_form.decorate(answers: answers)
+    @all_questions = award_form.steps.map(&:questions).flatten
+    @filled_answers = fetch_filled_answers
+
+    generate!
+  end
+
+  def generate!
+    main_header
+    render_data!
+  end
+end
