+++ app/views/admin/dashboard/index.html.slim
@@ -20,6 +20,7 @@
 
       - if policy(:report).show?
         = render "admin/feedbacks/pdf_reports"
+        = render "admin/reports/case_summaries"
 
       / TODO resources
       .well
